requirements.txt
cython
chardet
aiohttp
pyserial
pyyaml